--- src/ui/base/SvgIcon.ts
@@ -1,7 +1,0 @@
-import { FC, SVGProps } from "react";
-
-export type SvgIcon = FC<
-  SVGProps<SVGSVGElement> & {
-    title?: string | undefined;
-  }
->;
